refact : 디렉토리 구조 개선
