+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/generic_capsule/repository/ImageQueryRepository.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.capsule.repository;
+package site.timecapsulearchive.core.domain.capsule.generic_capsule.repository;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
