+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/generic_capsule/service/VideoService.java
@@ -1,20 +1,20 @@
-package site.timecapsulearchive.core.domain.capsule.service;
+package site.timecapsulearchive.core.domain.capsule.generic_capsule.service;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.timecapsulearchive.core.domain.capsule.entity.Capsule;
-import site.timecapsulearchive.core.domain.capsule.entity.Image;
-import site.timecapsulearchive.core.domain.capsule.repository.ImageQueryRepository;
+import site.timecapsulearchive.core.domain.capsule.entity.Video;
+import site.timecapsulearchive.core.domain.capsule.generic_capsule.repository.VideoQueryRepository;
 import site.timecapsulearchive.core.domain.member.entity.Member;
 import site.timecapsulearchive.core.infra.s3.data.dto.S3Directory;
 
 @Service
 @RequiredArgsConstructor
-public class ImageService {
+public class VideoService {
 
-    private final ImageQueryRepository imageQueryRepository;
+    private final VideoQueryRepository videoQueryRepository;
 
     @Transactional
     public void bulkSave(final List<String> fileNames, final Capsule capsule, final Member member) {
@@ -23,7 +23,7 @@ public class ImageService {
                 .map(fileName -> S3Directory.CAPSULE.generateFullPath(member.getId(), fileName))
                 .toList();
 
-            imageQueryRepository.bulkSave(Image.createOf(fullFileNames, capsule, member));
+            videoQueryRepository.bulkSave(Video.createOf(fullFileNames, capsule, member));
         }
     }
