+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/capsule/group_capsule/facade/GroupCapsuleFacade.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.capsule.facade;
+package site.timecapsulearchive.core.domain.capsule.group_capsule.facade;
 
 import lombok.RequiredArgsConstructor;
 import org.locationtech.jts.geom.Point;
@@ -7,8 +7,8 @@ import org.springframework.transaction.annotation.Transactional;
 import site.timecapsulearchive.core.domain.capsule.entity.Capsule;
 import site.timecapsulearchive.core.domain.capsule.group_capsule.data.dto.GroupCapsuleCreateRequestDto;
 import site.timecapsulearchive.core.domain.capsule.group_capsule.service.GroupCapsuleService;
-import site.timecapsulearchive.core.domain.capsule.service.ImageService;
-import site.timecapsulearchive.core.domain.capsule.service.VideoService;
+import site.timecapsulearchive.core.domain.capsule.generic_capsule.service.ImageService;
+import site.timecapsulearchive.core.domain.capsule.generic_capsule.service.VideoService;
 import site.timecapsulearchive.core.domain.capsuleskin.entity.CapsuleSkin;
 import site.timecapsulearchive.core.domain.capsuleskin.service.CapsuleSkinService;
 import site.timecapsulearchive.core.domain.group.entity.Group;
